Update main.py
Update to code, as I realized it was missing the updated image download code and details extraction code. I do want to say, this was made for an amazon business account!! It may not work with a regular amazon account, however, if you NEED a code for a regular amazon account, please get ahold of me on here. I will also add additional contact info to my BIO. This way you can also reach out with any questions. I will also be adding additional updates in the next few days to handle errors, and adding code to actually retrieve the images and embed them into the spreadsheet, as opposed to just adding the files paths! =D