Update main.py
Added more code comments, fixed some errors in code. If you have any suggestions or requests, feel free to contact me on facebook @Alysha Pursley or LinkedIn or wherever at same name! Working on making it actually add the images to the spreadsheet but I cant get the code correct for that for some reason and it breaks everything but as soon as I do I'll update this again! Happy coding!